Backend/requirements.txt
fastapi==0.104.1
pymongo==4.6.1
cryptography==41.0.5
python-jose[cryptography]==3.3.0
uvicorn==0.23.2
face_recognition==1.3.0
Pillow==10.4.0
numpy==1.26.4